add SwapCapacity (#7)
Allows callers to do other calculations and only set the capacity if the
old holds. Of course it won't consider whether other values held in the
calculations, but is better than nothing. Allowing size into it could be
possible if we traded those two atomics for a pointer atomic that held
both, but I'm not sure of the perf tradeoff there.