Add EvictSkip (#9)
Since the size is tracked in the regular cache, not the counting cache,
any evicted items that still have unreleased handles won't be included
in the size (after the evict) which means the cache will think it has
evicted enough.

This allows the counting cache the ability to have any of its items
skipped for evictions if they have any handles that aren't released.
I'll put a couple comment notes inline. The goal being to still have
that size counted, and thus other stuff evicted to keep the size from
ballooning.